feat: updated BE to support vercel
