+++ .gitignore
@@ -102,3 +102,4 @@ Thumbs.db
 
 # Test files (if any)
 test-*.js 
+.vercel
