+++ package.json
@@ -4,15 +4,16 @@
   "description": "AI-powered media plan generator for digital marketing campaigns",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js",
-    "dev": "nodemon server.js"
+    "dev": "node server.js"
   },
   "dependencies": {
     "@google/generative-ai": "^0.1.3",
     "cors": "^2.8.5",
     "dotenv": "^16.5.0",
     "express": "^4.18.2",
-    "helmet": "^7.1.0"
+    "helmet": "^7.1.0",
+    "serverless-http": "^3.2.0",
+    "vercel": "^43.2.0"
   },
   "devDependencies": {
     "nodemon": "^3.0.2"
