+++ vercel.json
@@ -1,0 +1,10 @@
+{
+  "version": 2,
+  "builds": [
+    { "src": "api/index.js", "use": "@vercel/node" }
+  ],
+  "routes": [
+    { "src": "/(.*)", "dest": "/public/$1" },
+    { "src": "/api/(.*)", "dest": "/api/index.js" }
+  ]
+}
